separete book detail control panel
